refatoring main module
