+++ lisapi/auth/__init__.py
@@ -1,5 +1,5 @@
 from flask import Blueprint
 
-errors = Blueprint('errors', __name__)
+auth = Blueprint('auth', __name__)
 
 from . import views
