+++ lisapi/pin/views.py
@@ -1,64 +1,14 @@
 from flask import (Blueprint, flash, jsonify, redirect, render_template,
                    request, url_for)
 from flask_login import current_user, login_required, login_user, logout_user
-from lisapi import db, login, socketio
+from lisapi import db
 from lisapi.helpers import helpers
-from lisapi.models.forms import ChangePassword, EditPin, Login, NewPin
+from lisapi.models.forms import EditPin, NewPin
 from lisapi.models.tables import Pin, User
-
-main = Blueprint('main', __name__)
-
-
-@socketio.on('updateStatus')
-def on_update(data):
-    """Update content in page, receive updateStatus and emit statusUpdated"""
-    data = helpers.statusInfo()
-    socketio.emit('statusUpdated', data.json)
+from . import pin
 
 
-@socketio.on('changeState')
-def change_pin_state(data):
-    """Change state for pins """
-    current_state = 0
-    pin_number = data['pin_number']
-    pin = Pin.query.filter_by(pin=pin_number).first()
-    pin_state = helpers.setPin(pin_number)
-    pin.state = pin_state
-
-    if pin_state is True:
-        current_state = 1
-
-    db.session.commit()
-
-    data = jsonify({
-        'pin': pin_number,
-        'currentState': current_state
-    })
-
-    socketio.emit('statusChanged', data.json)
-
-
-@socketio.on_error_default
-def error_handler(error):
-    socketio.emit('error', error)
-    print(e)
-
-
-@login.user_loader
-def load_user(id):
-    return User.query.get(id)
-
-
-@main.route('/', methods=['GET'])
-def dashboard():
-    context = {
-        'data': helpers.statusInfo(),
-        'active_menu': 'Dashboard'
-    }
-    return render_template('main/dashboard.html', **context)
-
-
-@main.route('/create', methods=['GET', 'POST'])
+@pin.route('/create', methods=['GET', 'POST'])
 def create_pin():
     form_new_pin = NewPin()
     if form_new_pin.validate_on_submit():
@@ -70,10 +20,10 @@ def create_pin():
         pin_exists = Pin.query.filter_by(pin=pin_number).first()
         if not check_pin:
             flash('Pin {} dont exists!'.format(pin_number), 'error')
-            return render_template('main/new.html', form=form_new_pin)
+            return render_template('pin/new.html', form=form_new_pin)
         elif pin_exists:
             flash('Pin {} not disponible!'.format(pin_number), 'error')
-            return render_template('main/new.html', form=form_new_pin)
+            return render_template('pin/new.html', form=form_new_pin)
         else:
             pin = Pin(name=pin_name, pin=pin_number,
                       color=pin_color, icon=pin_icon,
@@ -81,7 +31,7 @@ def create_pin():
             db.session.add(pin)
             db.session.commit()
             flash('Pin created!', 'success')
-            return redirect(url_for('main.list_pins'))
+            return redirect(url_for('pin.list_pins'))
 
     if len(form_new_pin.errors) > 0:
         flash('Check form data!', 'error')
@@ -91,10 +41,10 @@ def create_pin():
         'active_menu': 'New Pin'
     }
         
-    return render_template('main/new.html', **context)
+    return render_template('pin/new.html', **context)
 
 
-@main.route('/list', methods=['GET'])
+@pin.route('/list', methods=['GET'])
 @login_required
 def list_pins():
     list_pins = Pin.query.all()
@@ -104,10 +54,10 @@ def list_pins():
         'active_menu': 'List Pins'
     }
 
-    return render_template('main/list.html', **context)
+    return render_template('pin/list.html', **context)
 
 
-@main.route('/edit/<int:pin_id>', methods=['GET', 'POST'])
+@pin.route('/edit/<int:pin_id>', methods=['GET', 'POST'])
 @login_required
 def edit_pin(pin_id):
     form_editpin = EditPin()
@@ -120,7 +70,7 @@ def edit_pin(pin_id):
             pin.icon = form_editpin.icon.data
             db.session.commit()
             flash('Pin edited!', 'success')
-            return redirect(url_for('main.list_pins'))
+            return redirect(url_for('pin.list_pins'))
 
         if len(form_editpin.errors) > 0:
             flash('Check form data!', 'error')
@@ -134,20 +84,20 @@ def edit_pin(pin_id):
         'active_menu': 'Edit Pin'
     }
 
-    return render_template('main/edit.html', **context)
+    return render_template('pin/edit.html', **context)
 
 
-@main.route('/delete/<int:pin_id>', methods=['GET'])
+@pin.route('/delete/<int:pin_id>', methods=['GET'])
 @login_required
 def delete_pin(pin_id):
     pin = Pin.query.get(pin_id)
     db.session.delete(pin)
     db.session.commit()
     flash('Pin removed!', 'warning')
-    return redirect(url_for('main.list_pins'))
+    return redirect(url_for('pin.list_pins'))
 
 
-@main.route('/control', methods=['GET'])
+@pin.route('/control', methods=['GET'])
 @login_required
 def control_pins():
     list_pins = Pin.query.all()
@@ -157,9 +107,4 @@ def control_pins():
         'active_menu': 'Control Pins'
     }
 
-    return render_template('main/control.html', **context)
-
-
-@main.route('/about', methods=['GET'])
-def about():
-    return render_template('about.html', active_menu='About')
+    return render_template('pin/control.html', **context)
