+++ lisapi/templates/__sidemenu.html
@@ -14,13 +14,13 @@
       </a>
     </li>
     <li>
-      <a id="control_pins" href="{{ url_for('main.control_pins') }}">
+      <a id="control_pins" href="{{ url_for('pin.control_pins') }}">
         <i class="fas fa-cogs fa-lg"></i>
         Control Pins
       </a>
     </li>
     <li>
-      <a id="list_pins" href="{{ url_for('main.list_pins') }}">
+      <a id="list_pins" href="{{ url_for('pin.list_pins') }}">
         <i class="fas fa-edit fa-lg"></i>
         List Pins
       </a>
