+++ lisapi/templates/pin/control.html
@@ -83,7 +83,7 @@
   {% else %}
     <div class="center-align">
       <h4 class="mt2">Please register a pin.</h4>
-      <a class="btn mt2" href="{{ url_for('main.create_pin') }}">New Pin</a>
+      <a class="btn mt2" href="{{ url_for('pin.create_pin') }}">New Pin</a>
     </div>
   {% endif %}
   </div>
