+++ lisapi/templates/pin/edit.html
